Fix : 사장님 앱 retrofit api logger 추가, api 주소 변경 - 김성수 2025-02-20 03:46

+++ Qrscanner/app/src/main/java/com/example/qrscanner/ui/adapter/StoreAdapter.kt
@@ -16,7 +16,7 @@ class StoreAdapter(var storeList: List<StoreRes>, private val onBtnClicked: OnSt
     inner class StoreViewHolder(private val binding: StoreRowBinding) :
         RecyclerView.ViewHolder(binding.root) {
         fun bind(storeRes: StoreRes) {
-            val timestamp: Long = storeRes.orderDate.toLong()
+            val timestamp: Long = storeRes.orderDate
             val dateFormat = SimpleDateFormat("yyyy-MM-dd HH:mm:ss", Locale.getDefault())
             val dateString: String = dateFormat.format(Date(timestamp))
